+++ lib/core/src/treesitter/queries/rust/highlights.scm
@@ -1,145 +1,168 @@
-; Rust highlight queries
+; Rust highlight queries for tree-sitter-rust 0.24
+; Based on official tree-sitter-rust queries
 
-; Keywords
-[
-  "as"
-  "async"
-  "await"
-  "break"
-  "const"
-  "continue"
-  "crate"
-  "dyn"
-  "else"
-  "enum"
-  "extern"
-  "fn"
-  "for"
-  "if"
-  "impl"
-  "in"
-  "let"
-  "loop"
-  "match"
-  "mod"
-  "move"
-  "mut"
-  "pub"
-  "ref"
-  "return"
-  "self"
-  "Self"
-  "static"
-  "struct"
-  "super"
-  "trait"
-  "type"
-  "union"
-  "unsafe"
-  "use"
-  "where"
-  "while"
-] @keyword
+; Identifiers
 
-"fn" @keyword.function
-"return" @keyword.return
-["if" "else" "match"] @keyword.conditional
-["for" "while" "loop"] @keyword.repeat
-["use" "mod" "crate"] @keyword.import
-
-; Types
 (type_identifier) @type
 (primitive_type) @type.builtin
-(self_parameter (self) @type.builtin)
-
-; Functions
-(function_item name: (identifier) @function)
-(call_expression function: (identifier) @function)
-(call_expression function: (field_expression field: (field_identifier) @function.method))
-(macro_invocation macro: (identifier) @function.macro)
-(macro_invocation macro: (scoped_identifier name: (identifier) @function.macro))
-
-; Variables
-(identifier) @variable
-(parameter pattern: (identifier) @variable.parameter)
-(self_parameter (self) @variable.builtin)
-
-; Fields/Properties
 (field_identifier) @property
-(shorthand_field_initializer (identifier) @property)
 
-; Constants
-(const_item name: (identifier) @constant)
-(static_item name: (identifier) @constant)
+; Identifier conventions
+
+; Assume all-caps names are constants
 ((identifier) @constant
-  (#match? @constant "^[A-Z][A-Z0-9_]*$"))
+ (#match? @constant "^[A-Z][A-Z\\d_]+$"))
 
-; Literals
-(string_literal) @string
-(raw_string_literal) @string
-(char_literal) @string
-(escape_sequence) @string.escape
-(integer_literal) @number
-(float_literal) @number.float
-(boolean_literal) @boolean
+; Assume uppercase names are constructors
+((identifier) @constructor
+ (#match? @constructor "^[A-Z]"))
+
+; Assume that uppercase names in paths are types
+((scoped_identifier
+  path: (identifier) @type)
+ (#match? @type "^[A-Z]"))
+((scoped_type_identifier
+  path: (identifier) @type)
+ (#match? @type "^[A-Z]"))
+
+; Function calls
+
+(call_expression
+  function: (identifier) @function)
+(call_expression
+  function: (field_expression
+    field: (field_identifier) @function.method))
+(call_expression
+  function: (scoped_identifier
+    "::"
+    name: (identifier) @function))
+
+(generic_function
+  function: (identifier) @function)
+(generic_function
+  function: (scoped_identifier
+    name: (identifier) @function))
+(generic_function
+  function: (field_expression
+    field: (field_identifier) @function.method))
+
+(macro_invocation
+  macro: (identifier) @function.macro
+  "!" @function.macro)
+(macro_invocation
+  macro: (scoped_identifier
+    name: (identifier) @function.macro)
+  "!" @function.macro)
+
+; Function definitions
+
+(function_item (identifier) @function)
+(function_signature_item (identifier) @function)
 
 ; Comments
+
 (line_comment) @comment
 (block_comment) @comment
-((line_comment) @comment.documentation
-  (#match? @comment.documentation "^///"))
-((block_comment) @comment.documentation
-  (#match? @comment.documentation "^/\\*\\*"))
+(line_comment (doc_comment)) @comment.documentation
+(block_comment (doc_comment)) @comment.documentation
+
+; Punctuation
+
+"(" @punctuation.bracket
+")" @punctuation.bracket
+"[" @punctuation.bracket
+"]" @punctuation.bracket
+"{" @punctuation.bracket
+"}" @punctuation.bracket
+
+(type_arguments
+  "<" @punctuation.bracket
+  ">" @punctuation.bracket)
+(type_parameters
+  "<" @punctuation.bracket
+  ">" @punctuation.bracket)
+
+"::" @punctuation.delimiter
+":" @punctuation.delimiter
+"." @punctuation.delimiter
+"," @punctuation.delimiter
+";" @punctuation.delimiter
+
+; Parameters and variables
+
+(parameter (identifier) @variable.parameter)
+(identifier) @variable
+
+; Lifetimes
+
+(lifetime (identifier) @label)
+
+; Keywords
+
+"as" @keyword
+"async" @keyword
+"await" @keyword
+"break" @keyword
+"const" @keyword
+"continue" @keyword
+"default" @keyword
+"dyn" @keyword
+"else" @keyword
+"enum" @keyword
+"extern" @keyword
+"fn" @keyword
+"for" @keyword
+"gen" @keyword
+"if" @keyword
+"impl" @keyword
+"in" @keyword
+"let" @keyword
+"loop" @keyword
+"macro_rules!" @keyword
+"match" @keyword
+"mod" @keyword
+"move" @keyword
+"pub" @keyword
+"raw" @keyword
+"ref" @keyword
+"return" @keyword
+"static" @keyword
+"struct" @keyword
+"trait" @keyword
+"type" @keyword
+"union" @keyword
+"unsafe" @keyword
+"use" @keyword
+"where" @keyword
+"while" @keyword
+"yield" @keyword
+
+; Named keyword nodes
+(crate) @keyword
+(mutable_specifier) @keyword
+(super) @keyword
+(use_list (self) @keyword)
+(scoped_use_list (self) @keyword)
+(scoped_identifier (self) @keyword)
+(self) @variable.builtin
+
+; Literals
+
+(char_literal) @string
+(string_literal) @string
+(raw_string_literal) @string
+(boolean_literal) @boolean
+(integer_literal) @number
+(float_literal) @number
+(escape_sequence) @string.escape
 
 ; Attributes
+
 (attribute_item) @attribute
 (inner_attribute_item) @attribute
 
 ; Operators
-[
-  "+"
-  "-"
-  "*"
-  "/"
-  "%"
-  "="
-  "=="
-  "!="
-  "<"
-  ">"
-  "<="
-  ">="
-  "&&"
-  "||"
-  "!"
-  "&"
-  "|"
-  "^"
-  "~"
-  "<<"
-  ">>"
-  "+="
-  "-="
-  "*="
-  "/="
-  "%="
-  "&="
-  "|="
-  "^="
-  "<<="
-  ">>="
-  ".."
-  "..="
-  "=>"
-  "->"
-  "::"
-  "?"
-] @operator
 
-; Punctuation
-["(" ")" "[" "]" "{" "}"] @punctuation.bracket
-[";" "," ":" "."] @punctuation.delimiter
-["#" "!"] @punctuation.special
-
-; Lifetime
-(lifetime) @label
+"*" @operator
+"&" @operator
+"'" @operator